backend/profile/requirements.txt
psycopg2-binary
PyJWT
bcrypt
boto3
